Fix install requirements for chapter 7
Previously, requirements would be an array when is_chapter7 was true.
In line 7, we were passing an array to subprocess.run. However, given
that requirements was an array, it would be an array within an array

The code was:
process_install = subprocess.run(
  ["python", "-m", "pip", "install", "-r", requirements], stdout=subprocess.PIPE, stderr=subprocess.PIPE
)

So, by requirements being an array, it would end up executing this:
process_install = subprocess.run(
  ["python", "-m", "pip", "install", "-r", ["requirements-chapter7.txt", "-f", "https://download.pytorch.org/whl/torch_stable.html"]], stdout=subprocess.PIPE, stderr=subprocess.PIPE
)

And thus it was erroring out with this message:
TypeError: expected str, bytes or os.PathLike object, not list